refactor: creating folder to Chat in components and changing name of archive
